tests/wpt/web-platform-tests/preload/link-header-on-subresource.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/preload/resources/preload_helper.js"></script>
<script>
    var t = async_test('Makes sure that Link headers on subresources preload resources');
</script>
<link rel=stylesheet href="resources/dummy-preloads-subresource.css?link-header-on-subresource">
<script src="resources/dummy.js?pipe=trickle(d5)&link-header-on-subresources"></script>
<script>
    window.addEventListener("load", t.step_func(function() {
        verifyPreloadAndRTSupport();
        verifyNumberOfDownloads("/preload/resources/CanvasTest.ttf?link-header-on-subresource", 1);
        t.done();
    }));
</script>
